Fix CRC calculation of UDI images on LP64 platforms
for example most Linux 64 bits. The reference algorithm comes from the
32 bits era and relies on a "long" type of 32 bits. That only works on
Windows 32 bits (ILP32), Windows 64 bits (LLP64) or Linux 32 bits
(LLP64).

+++ src/types/udi.cpp
@@ -27,7 +27,7 @@ typedef struct
 
 static uint32_t crc32 (const uint8_t *buf, int len)
 {
-	long crc = ~0;
+	int32_t crc = ~0;
 	for (int i = 0; i < len; i++)
 	{
 		crc ^= ~buf[i];
